requirements.txt
# Google Sheets API
google-api-python-client>=2.0.0
google-auth>=2.0.0
google-auth-oauthlib>=0.5.0
google-auth-httplib2>=0.1.0

# HyperFormula for local formula evaluation
hyperformula>=3.0.0

# Configuration management
python-dotenv>=1.0.0

# Utilities
pandas>=1.3.0
openpyxl>=3.0.0
